get method for our deliverables

+++ DTID.BusinessLogic/Models/Wage.cs
@@ -8,7 +8,7 @@ namespace DTID.BusinessLogic.Models
     {
         public int ID { get; set; }
         public Year Year { get; set; }
-        public int Wages { get; set; }
+        public double Wages { get; set; }
         public DateTime DateCreated { get; set; }
         public DateTime DateUpdated { get; set; }
     }
